+++ curatorx/research/__init__.py
@@ -1,0 +1,1 @@
+"""Safe, configured external media research services."""